Added some changes

+++ internal/repository/MetricsRepository.go
@@ -1,12 +1,12 @@
 package repository
 
 type MetricsRepository interface {
-	SaveCounter(name string, value float64) (float64, error)
-	SaveGauge(name string, value int64) (int64, error)
+	SaveCounter(name string, value int64) (int64, error)
+	SaveGauge(name string, value float64) (float64, error)
 
-	GetCounter(name string) (float64, error)
+	GetCounter(name string) (int64, error)
 	GetGauge(name string) (float64, error)
 
-	GetAllCounters() (map[string]float64, error)
-	GetAllGauges() (map[string]int64, error)
+	GetAllCounters() (map[string]int64, error)
+	GetAllGauges() (map[string]float64, error)
 }
